Remove QSA flag
This would combine the original query string with the query string in our rewrite rule, and causes the query string parameters to be reordered. However there is no query string in our new URL, so there is nothing to combine it with and it just reorders the original query parameters.